moved findBinary function and the required interface to shell.ts
